Add logo to the header, and simplify layout inheritance (#4)
`govuk-eleventy-plugin` now allows you to customise the full contents of the header (logo and organisation name included). So I am loading in the SVG for the logo, and using that string for the new `organisationLogo` option. 

That means the only thing that needs to be overridden is the footer, and as as we’re only using one layout at the moment, have simplified things a bit to inherit from the plug-ins’ `product` layout and only extend only the `footer` block.